pyproject.toml
[project]
name = "db-chatagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "openai-agents>=0.6.2",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
